tweaks to pack duplication merger
